proto2js: allow package name override
proto2js allows users to specify the name of the resulting class.
However, it looks for a package definition inside the AST of the
parsed .proto file and exits with failure in case the one specified
on the command line is different, otherwise it will just pick up
whatever is declared in the .proto file or nothing.

This is in contrast with proto2js help string which explicitly allows
users to specify a new package name on the command line.

This commit rectifies the situation. In case a new package name is
specified, it is injected into the AST. This has the side-effect
that it is effectively possible to override whatever package
declaration was present in the .proto file.