Declare namespaces and syntax in pbjs json output
This patch declares namespaces in json output by introducing the
isNamespace flag. This is backwards compatible as the previous errant
behavior would instantiate messages for namespaces. This will still
happen if an old version of pbjs is used, but this works fine, anyway.

Syntax fields are also set in the output.

Fixes #482 and #469