Use same node when shelling out to jsdoc
Right now we assume that the node in the process.env['PATH'] will be the right one. If the user uses a certain version of node to invoke pbts (or doesn't have node on their PATH) then this does the wrong thing. We should spawn another copy of the same node binary that's already running.