feat: Add the `--sound-types` flag
This change adds a flag to enforce the generation of sound type
annotations (which aid in the generation of Flow types, #939).

TypeScript allows certain [unsound
type](https://www.typescriptlang.org/docs/handbook/type-compatibility.html#a-note-on-soundness)
annotations as valid, for compatibility purposes with pre-existing
JavaScript code. That's completely fine since it still allows code to be
safer than what it would be without the type annotations. However, Flow
was designed with a different set of trade-offs, erring on the side of
[soudnness](https://flow.org/en/docs/lang/types-and-expressions/).

With this flag passed in, the generated code will:

a) No longer establish any inheritance between the interface
   (properties) and message types. Interface types allow every single
   property to be optional (having `undefined` as a valid value),
   whereas message types do not allow this. Thus, it would be a
   violation of the [Liskov substitution
   principle](https://en.wikipedia.org/wiki/Liskov_substitution_principle),
   and therefore Flow does not allow this inheritance. Instead, all the
   fields in a interface type explicitly allow both interface and
   message types to be passed in instead of just interface types.
b) When the `--force-message` flag is also passed, have the message
   objects always forward calls to `.create()` to `.fromObject()` instead
   of the constructor. The reason for this is that the properties may
   have fields that are valid interface types, but the message type
   expects these to be message types.