feat: support $unknownFields
This add `$unknownFields` property to generated code. It is populated on
`.decode()` when decoder encounters currently unknown field, and
similarly used by encoder to serialize unknown fields as they are into a
binary data.

If user stores unknown fields and then upgrades their protobufs to the
latest spec - there is a chance that some of the previously unknown
fields will become known. This is handled in the encoder by attempting
to decode the fields first.

The name of the field starts with `$` because it is not a valid
character for a field name according to protobuf v3 specification.