Fix line accounting for `alternateCommentMode`
When `alternateCommentMode` is enabled, the tokenizer treats a block of multiple end-of-line (`//`) comments as a single line when keeping track of line numbers, making error messages very difficult to understand, since they refer to the wrong line number.

This change increments the line number in the inner loop, so that each line is accounted correctly.