fix: invalid syntax in descriptor.proto
A recent change (#2075) edited descriptor.proto to include a `reserved`
statement that combines a field number and field name in the same line.
This is specifically called out in the proto language guide as invalid.

https://protobuf.dev/programming-guides/editions/#reserved

> Note that you can’t mix field names and numeric values in the same
> `reserved` statement.

Protobuf.js itself seems to handle this syntax just fine, but other
tools like protoc will choke on it.